add student documents model

+++ src/main/java/vp/spring/rcs/model/Subject.java
@@ -9,13 +9,13 @@ public class Subject {
 	
 	@Id
     @GeneratedValue
-    private int id;
+    private Long id;
 	
 	private String name;
 	
 	private String description;
 	
-	public Subject(int id, String name, String description) {
+	public Subject(Long id, String name, String description) {
 		super();
 		this.id = id;
 		this.name = name;
@@ -26,11 +26,11 @@ public class Subject {
 		
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
